rename tools, workflows, and agents for clarity
